ajouterPosition
